// movin/src/components/units/boardComment/Routing_Comment/BoardCommentWrite.types.ts
import { ChangeEvent, Dispatch, SetStateAction } from "react";

export interface IBoardCommentWriteUIProps {
  onChangeWriter: (event: ChangeEvent<HTMLInputElement>) => void;
  onChangePassword: (event: ChangeEvent<HTMLInputElement>) => void;
  onChangeContents: (event: ChangeEvent<HTMLTextAreaElement>) => void;
  onClickWrite: () => void;
  contents: string;
  setStar: Dispatch<SetStateAction<number>>;
  el: any;
  isEdit: any;
  setEditId: any;
}

export interface IUpdateBoardCommentInput {
  contents?: string;
  rating?: number;
}